% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{energy_weekly}
\alias{energy_weekly}
\title{Weekly petroleum and gas consumption}
\format{Data frame with 6 columns
\describe{
\item{EconYear}{Year of the economic quarter.  Because economic quarters are
deemed to begin and end on week boundaries, this may not correspond exactly
to the calendar year for the data.}
\item{quarter}{Economic quarter.}
\item{week}{Week within the quarter.  The number of weeks in a quarter varies
a bit (usually it's 13 or 14).  Only weeks that exist for a quarter are included in the
dataset.}
\item{petrol}{Petroleum product consumption for the week.  Given as a
standardized value.}
\item{gas}{Natural gas consumption in certain product categories (see
details) for the week.  Given as a standardized value.}
\item{dataset}{Dataset the quarter was originally assigned to (training, dev,
or test)}
}}
\description{
This data is organized by economic year, quarter and week.  All values have
been standardized.
}
\details{
Natural gas consumption is given only for the categories
'Industrial Consumption', 'Lease and Plant Fuel Consumption', and 'Pipeline &
Distribution Use'.  Specifically, natural gas for electrical generation is
excluded because it would be redundant with the electicity data, and
residential use is excluded because it is primarily temperature driven.
}
