% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{gdp_quarterly}
\alias{gdp_quarterly}
\title{Quarterly GDP}
\format{Data frame with 4 columns
\describe{
\item{EconYear}{Year of the economic quarter.  Because economic quarters are
deemed to begin and end on week boundaries, this may not correspond exactly
to the calendar year for the data.}
\item{quarter}{Economic quarter.}
\item{gdp}{US Gross Domestic Product, billions of USD}
\item{dataset}{Dataset the quarter was originally assigned to (training, dev,
or test)}
}}
\description{
The data is organized by economic year and quarter.  GDP values are in
billions of US dollars (inflation adjusted or not?)
}
