% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temploadmod.R
\name{load_model}
\alias{load_model}
\title{Create models for load as a function of temperature}
\usage{
load_model(model = stats::lm, hourly_data = NULL, ...)
}
\arguments{
\item{model}{Modeling function to use.  The function must conform to the
\code{\link[stats]{lm}} conventions.}

\item{hourly_data}{Data frame of hourly data.  If omitted, then
\code{\link{region_hourly}} is used.}

\item{...}{Additional arguments passed to the modeling function.}
}
\value{
List of model objects, one for each region.
}
\description{
(Note this is creating a model for a variable called "load", not, as the name
might suggest, loading a model from some storage location.)
The modeling is done separately for each region, and a list of regional
models is returned.
}
