% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{region_hourly}
\alias{region_hourly}
\title{Hourly load and temperature data by region}
\format{Data frame with 7 columns
\describe{
\item{EconYear}{Year of the economic quarter.  Because economic quarters are
deemed to begin and end on week boundaries, this may not correspond exactly
to the calendar year for the data.}
\item{quarter}{Economic quarter.}
\item{week}{Week within the quarter.}
\item{region}{NERC region.}
\item{time}{Date and hour.}
\item{load}{Electrical load reported in the region (MW).}
\item{temperature}{Average temperature across the region.}
}

The data also has a \code{notes} attribute that may provide additional
relevant information.}
\description{
This data frame contains all of the data available at hourly resolution,
namely, load and temperature.  These items are tabulated by time and NERC
region.
}
