% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{residual_load}
\alias{residual_load}
\title{Hourly residual load by region}
\format{Data frame with 6 columns
\describe{
\item{EconYear}{Economic year.  Since quarters are forced to begin and
end on week boundaries, this may not match the calendar year.}
\item{quarter}{Economic quarter (1-4).}
\item{week}{Week within the quarter.}
\item{region}{NERC region.}
\item{time}{Date and hour.}
\item{residual_load}{Electrical load minus the load predicted by the temperature-load model.}
}}
\description{
This data frame contains the hourly load data, with load values predicted
from the temperature data subtracted off.
}
