% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize one or more variables in a data frame}
\usage{
standardize(df, vars)
}
\arguments{
\item{df}{Data frame with the data to be standardized}

\item{vars}{Vector of names of columns to be standardized.}
}
\value{
Data frame with the requested variables standardized.
}
\description{
Standardize by subtracting off the mean and dividing by the standard
deviation.
}
\details{
The mean and scale are recorded in named vectors attached to the
output data frame as attributes names \code{mean} and \code{scale},
respectively.
}
\keyword{internal}
