% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize_byrgn}
\alias{standardize_byrgn}
\title{Standardize variables by region}
\usage{
standardize_byrgn(df, vars)
}
\arguments{
\item{df}{Data frame with the data to be standardized.}

\item{vars}{Vecotr of names of columns to be standardized.}
}
\value{
Data frame with the requested variables standardized.
}
\description{
This function applies the standardize function to each regional grouping and
recombines the resutls into a single data frame.
}
\details{
The means and scales for the transformations are stored in the mean and scale
attributes of the output.  Each is a list indexed by region, with each list
element being a named vector indexed by variable.
}
