% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time.R
\name{timesince}
\alias{timesince}
\title{Return time since the earliest time in a vector}
\usage{
timesince(t, unit = "hours")
}
\arguments{
\item{t}{Vector of times.}

\item{unit}{Units to return the differences in.}
}
\description{
For a vector of times, return a vector of times since the minimum
time in the vector.
}
