% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{unstandardize}
\alias{unstandardize}
\title{Unstandardize variables in a data frame}
\usage{
unstandardize(df)
}
\arguments{
\item{df}{Data frame with data to be unstandardized.}
}
\value{
Data frame of unstandardized variables
}
\description{
Unstandardize by multiplying by the scale and adding the mean.
}
\details{
The means and scales for any standardized variables must be recorded in the
\code{mean} and \code{scale} attributes of the input data frame.  Only
variables mentioned in those attributes will be modified.
}
\keyword{internal}
