% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{unstandardize_byrgn}
\alias{unstandardize_byrgn}
\title{Unstandardize variables by region}
\usage{
unstandardize_byrgn(df)
}
\arguments{
\item{df}{Data frame with the data to be unstandardized}
}
\value{
Data frame of unstandardized variables
}
\description{
This function reverses the \code{\link{standardize_byrgn}} function.
}
