% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{weekcount}
\alias{weekcount}
\title{Number of weeks in each economic quarter}
\format{A Data frame with three columns
\describe{
\item{EconYear}{Economic year.}
\item{quarter}{Economic quarter.}
\item{n}{Number of weeks in the quarter.}}}
\description{
This table gives a count of the number of weeks in each quarter.  Most quarters
have 13 weeks (91 days).  The actual number of days in a quarter varies from 90
days (Q1 in non-leap years) to 92 (Q3 and Q4).  Note, however, that even when the
number of days matches, the actual days being counted may vary a bit, with extra
days picked up at one end of the quarter and dropped at the other.
}
\details{
There are four weeks in the dataset with a week count other than 13.  2009-Q4, 2010-Q3,
and 2015-Q4 each have 14 weeks, while 2010-Q1 has only 12.
}
